\name{clio_infra-package}
\alias{clio_infra-package}
\alias{clio_infra}
\docType{package}
\title{
This package provides functions to conveniently transform clio-infra data into different formats.
}
\description{
The clio-infra project provides important indicators on economic and social history of the world for the past 5 centuries. This package provides a number of functions (currently, version 0.1, only 1) to handle the clio-infra data}
\details{
\tabular{ll}{
Package: \tab clio_infra\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-05-12\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Richard Zijdeman

Maintainer: Richard Zijdeman (richard.zijdeman@iisg.nl)
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
